+++ src/app/np-tree-view-demo/np-tree-view-demo.component.ts
@@ -28,7 +28,7 @@ export class NpTreeViewDemoComponent implements OnInit {
   }),
   ...
 ];`;
-  itemTemplateText = `<ng-template #nodeTemplate let-item="item">
+  itemTemplateText = `<ng-template #itemTemplate let-item="item">
   {{item.label}}, {{item.key}}
 </ng-template>`;
 
@@ -116,41 +116,41 @@ export class NpTreeViewDemoComponent implements OnInit {
   items3: NpTreeViewItem[] = [
     new NpTreeViewItem({
       label: 'Item 1',
-      key: 'Item 1',
+      data: { id: '1' },
       childItems: [
         new NpTreeViewItem({
-          label: 'Item 1.1', key: 'Item 1.1', childItems: [
-            new NpTreeViewItem({ label: 'Item 1.1.1', key: 'Item 1.1.1' }),
-            new NpTreeViewItem({ label: 'Item 1.1.2', key: 'Item 1.1.2' }),
-            new NpTreeViewItem({ label: 'Item 1.1.3', key: 'Item 1.1.3' }),
-            new NpTreeViewItem({ label: 'Item 1.1.4', key: 'Item 1.1.4' }),
+          label: 'Item 1.1', data: { id: '1.1' }, childItems: [
+            new NpTreeViewItem({ label: 'Item 1.1.1', data: { id: '1.1.1' } }),
+            new NpTreeViewItem({ label: 'Item 1.1.2', data: { id: '1.1.2' } }),
+            new NpTreeViewItem({ label: 'Item 1.1.3', data: { id: '1.1.3' } }),
+            new NpTreeViewItem({ label: 'Item 1.1.4', data: { id: '1.1.4' } }),
           ]
         }),
-        new NpTreeViewItem({ label: 'Item 1.2', key: 'Item 1.2' }),
-        new NpTreeViewItem({ label: 'Item 1.3', key: 'Item 1.3' }),
-        new NpTreeViewItem({ label: 'Item 1.4', key: 'Item 1.4' }),
+        new NpTreeViewItem({ label: 'Item 1.2', data: { id: '1.2' } }),
+        new NpTreeViewItem({ label: 'Item 1.3', data: { id: '1.3' } }),
+        new NpTreeViewItem({ label: 'Item 1.4', data: { id: '1.4' } }),
       ]
     }),
     new NpTreeViewItem({
       label: 'Item 2',
-      key: 'Item 2',
+      data: { id: '2' },
       childItems: [
         new NpTreeViewItem({
-          label: 'Item 2.1', key: 'Item 2.1', childItems: [
-            new NpTreeViewItem({ label: 'Item 2.1.1', key: 'Item 2.1.1' }),
-            new NpTreeViewItem({ label: 'Item 2.1.2', key: 'Item 2.1.2' }),
-            new NpTreeViewItem({ label: 'Item 2.1.3', key: 'Item 2.1.3' }),
-            new NpTreeViewItem({ label: 'Item 2.1.4', key: 'Item 2.1.4' }),
+          label: 'Item 2.1', data: { id: '2.1' }, childItems: [
+            new NpTreeViewItem({ label: 'Item 2.1.1', data: { id: '2.2.1' } }),
+            new NpTreeViewItem({ label: 'Item 2.1.2', data: { id: '2.2.2' } }),
+            new NpTreeViewItem({ label: 'Item 2.1.3', data: { id: '2.2.3' } }),
+            new NpTreeViewItem({ label: 'Item 2.1.4', data: { id: '2.2.4' } }),
           ]
         }),
-        new NpTreeViewItem({ label: 'Item 2.2', key: 'Item 2.2' }),
-        new NpTreeViewItem({ label: 'Item 2.3', key: 'Item 2.3' }),
-        new NpTreeViewItem({ label: 'Item 2.4', key: 'Item 2.4' }),
+        new NpTreeViewItem({ label: 'Item 2.2', data: { id: '2.2' } }),
+        new NpTreeViewItem({ label: 'Item 2.3', data: { id: '2.3' } }),
+        new NpTreeViewItem({ label: 'Item 2.4', data: { id: '2.4' } }),
       ]
     }),
-    new NpTreeViewItem({ label: 'Item 3', key: 'Item 3', childItems: [] }),
-    new NpTreeViewItem({ label: 'Item 4', key: 'Item 4' }),
-    new NpTreeViewItem({ label: 'Item 5', key: 'Item 5' }),
+    new NpTreeViewItem({ label: 'Item 3', data: { id: '3' }, childItems: [] }),
+    new NpTreeViewItem({ label: 'Item 4', data: { id: '4' } }),
+    new NpTreeViewItem({ label: 'Item 5', data: { id: '5' } }),
   ];
   selection3 = [{ label: 'Item 4', key: 'Item 4' }, { label: 'Item 1.1.1', key: 'Item 1.1.1' }];